+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/scanner_helpers.py
@@ -278,7 +278,7 @@ class ScanStatus(AbstractScanStatus):
 
         get_latest_by = "scan_tag__time"
 
-        permissions = [("can_export_completed_scans", _("Can export history of completed scans"))]
+        permissions = [("export_completed_scanstatus", _("Can export history of completed scans"))]
 
     def __str__(self):
         return f"{self.scanner}: {self.start_time}"